webapp/requirements.txt
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
django-debug-toolbar==4.2.0
django-libsass==0.9
fontawesomefree==6.5.1
python-dotenv==1.0.1
django-widget-tweaks
gunicorn
django-countries@git+https://github.com/SmileyChris/django-countries@58f258402072a18756a15daa81325428382bd946
django-environ
google-cloud-secret-manager
django-storages[google]
google-cloud-storage
django-select2
django-guest-user
django-tomselect
pillow
psycopg2-binary
pyyaml
django-addanother
